+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## Next Release
 ###Internal
 - Migrated `LoginPinScreen` to Mobius
-- [In Progress: 24th Jul 2020] Extract facility selection UI into a separate component
-- [In Progress: 28th Jul 2020] Migrate `ForgotPinCreateNewPinScreen` to Mobius
+- Extracted facility selection UI into a separate component
 - [In Progress: 28th Jul 2020] Migrate `OverdueScreen` to Mobius
+- [In Progress: 28th Jul 2020] Migrate `ForgotPinCreateNewPinScreen` to Mobius
 
 ## On Demo
 ### Features